# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HRMS API
  version: 1.49.0
  description: HR Management System — Phase 1
paths:
...
  /api/v1/onboarding/:
    get:
      operationId: onboarding_list
      description: |-
        Employee onboarding checklists.

        Read gated on onboarding:read; create/toggle on onboarding:write.
        Creating a checklist seeds the DEFAULT_ITEMS template.
      tags:
      - onboarding
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnboardingChecklist'
          description: ''
    post:
      operationId: onboarding_create
      description: |-
        Employee onboarding checklists.

        Read gated on onboarding:read; create/toggle on onboarding:write.
        Creating a checklist seeds the DEFAULT_ITEMS template.
      tags:
      - onboarding
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingChecklistRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OnboardingChecklistRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingChecklist'
          description: ''
  /api/v1/onboarding/{id}/:
    get:
      operationId: onboarding_retrieve
      description: |-
        Employee onboarding checklists.

        Read gated on onboarding:read; create/toggle on onboarding:write.
        Creating a checklist seeds the DEFAULT_ITEMS template.
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - onboarding
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingChecklist'
          description: ''
    put:
      operationId: onboarding_update
      description: |-
        Employee onboarding checklists.

        Read gated on onboarding:read; create/toggle on onboarding:write.
        Creating a checklist seeds the DEFAULT_ITEMS template.
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - onboarding
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingChecklistRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/OnboardingChecklistRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingChecklist'
          description: ''
    patch:
      operationId: onboarding_partial_update
      description: |-
        Employee onboarding checklists.

        Read gated on onboarding:read; create/toggle on onboarding:write.
        Creating a checklist seeds the DEFAULT_ITEMS template.
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - onboarding
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedOnboardingChecklistRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedOnboardingChecklistRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingChecklist'
          description: ''
    delete:
      operationId: onboarding_destroy
      description: |-
        Employee onboarding checklists.

        Read gated on onboarding:read; create/toggle on onboarding:write.
        Creating a checklist seeds the DEFAULT_ITEMS template.
      parameters:
      - in: path
        name: id
        schema:
          type: string
        required: true
      tags:
      - onboarding
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/roles/{code}/reset-to-defaults/:
    post:
      operationId: roles_reset_to_defaults_create
      description: POST /api/v1/roles/{code}/reset-to-defaults/
      parameters:
      - in: path
        name: code
        schema:
          type: string
        required: true
      tags:
      - roles
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
  schemas:
...
    OnboardingChecklist:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        employee_id:
          type: string
          format: uuid
        status:
          allOf:
          - $ref: '#/components/schemas/OnboardingChecklistStatusEnum'
          readOnly: true
        started_at:
          type: string
          format: date-time
          readOnly: true
        completed_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/OnboardingItem'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - completed_at
      - created_at
      - employee_id
      - id
      - items
      - started_at
      - status
    OnboardingChecklistRequest:
      type: object
      properties:
        employee_id:
          type: string
          format: uuid
      required:
      - employee_id
    OnboardingChecklistStatusEnum:
      enum:
      - in_progress
      - completed
      type: string
      description: |-
        * `in_progress` - In progress
        * `completed` - Completed
    OnboardingItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        label:
          type: string
          maxLength: 200
        done:
          type: boolean
        order:
          type: integer
          maximum: 32767
          minimum: 0
      required:
      - id
      - label
...
    PatchedOnboardingChecklistRequest:
      type: object
      properties:
        employee_id:
          type: string
          format: uuid
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT